requirements.txt
discord.py==2.3.2
asyncio==3.4.3
dotenv==1.0.0